Merged SFX and Music into the same file
	Originally, I planned on extending either music to sfx or sfx to music,
however, since they have basically the same code I merged them and split
their functions up to make the files a bit neater. SFX and music work
the same way, the only difference is the pause/mute. Pausing the song
stops it and unmuting it plays it where it originally stopped. Muting
the sfx mutes ALL sfx so keep that in mind.