Simple changes to the the level of access to musicAndSFX.
	Changed most functions to static so that it could pause/play the music
by just calling the class. This is not perfect, however, in this case,
since we only use one background song then it is implied that the
"play/pause" music function refers to the only background song. On the
other hand, it is necessary to make objects for SFX because there are
multiple audio clips.